requirements.txt
#langchain
#PyPDF2
python-dotenv
streamlit==1.28.0
streamlit-agraph
edgedb
#faiss-cpu
#weaviate-client
#altair
#tiktoken
#comet_llm
plotly
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2